+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 予約検索にcountDocuments指定を追加
 - 予約にuseActionExists属性を追加
 - 予約検索に$projectionを指定できるように調整
+- 予約使用時にagent.identifierを指定できるように調整
 
 ### Deprecated